moved listingdetails screen to screens folder
