+++ app/navigation/FeedNavigator.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+import ListingsScreen from "../screens/ListingsScreen";
+import ListingsDetailScreen from "../screens/ListingsDetailScreen";
+
+const Stack = createStackNavigator();
+
+const FeedNavigator = () => (
+  <Stack.Navigator>
+    <Stack.Screen name="Listings" component={ListingsScreen} />
+    <Stack.Screen name="ListingDetails" component={ListingsDetailScreen} />
+  </Stack.Navigator>
+);
+
+export default FeedNavigator;
