+++ app/screens/ListingsDetailScreen.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { View, StyleSheet, Image } from "react-native";
 
-import AppText from "./AppText";
+import AppText from "../components/AppText";
 import colors from "../config/colors";
-import ListItem from "./ListItem";
+import ListItem from "../components/ListItem";
 
 function ListingsDetailScreen() {
   return (
